Created getValue method
getValue method returns value only if the item's age has passed maturaton age. The specs also suggest this occurs only for food items, so I created an if statement that checks the instance of the item. Working on implementing that within the Food class (creating a constructor) but unsure at this time.