Fixed all issues in step two, time to finish step three
Had an issue where the age was not the same between original and copy of Item. This resulted in a false output when comparing the two, so I added setAge to each if statement to ensure it was the same. Also had an issue in the Field class where I was iterating over height twice, obviously this does not work when the height and weight are different