Started individual functionality of scarecow
Created Scarecrow class. This function works with a Crow that eats the crop, and a Scarecrow scares Crows away. Currently unsure whether to inherit it from Item or make it its own, as it does use a lot of the functionality of Item. Crow may need to be by itself or belong to the Farm itself?