Deleted Crow, stuck with decision to make Scarecrow an Item subclass
So, I deleted crow because the functionality is fairly simple and can be encapsulated in a method in Field, which I haven't started writing yet. For the sake of both logic and ease of implementation, Scarecrow will remain a subclass of Item, so I can utilise Item methods. In Farm, I added an if statement where you cannot plant if soil is untilled. Soil must be tilled to plant (but I can't find that specifically mentioned in the assignment so it might be up to interpretation)