Removed placeScarecrow from Item parent class
Made a silly mistake. First, in Farm class, I was creating an instance of Scarecrow as an Item. The problem with this is it meant I needed to define it in the Item class. So I needed to "override" it in the Scarecrow class. Except, I realised I could just make an instance of Scarecrow and have now deleted the method in Item